+++ src/main/kotlin/io/verik/intellij/inspection/common/AbstractVerikInspection.kt
@@ -16,7 +16,11 @@
 
 package io.verik.intellij.inspection.common
 
+import com.intellij.codeInspection.ProblemsHolder
+import com.intellij.psi.PsiElementVisitor
 import org.jetbrains.kotlin.idea.inspections.AbstractKotlinInspection
+import org.jetbrains.kotlin.psi.KtFile
+import org.jetbrains.kotlin.psi.KtVisitorVoid
 
 abstract class AbstractVerikInspection : AbstractKotlinInspection() {
 
@@ -26,5 +30,27 @@ abstract class AbstractVerikInspection : AbstractKotlinInspection() {
 
     override fun isEnabledByDefault(): Boolean {
         return true
+    }
+
+    final override fun buildVisitor(holder: ProblemsHolder, isOnTheFly: Boolean): PsiElementVisitor {
+        return if (isEnabled(holder)) {
+            buildVisitor(holder)
+        } else NULL_VISITOR
+    }
+
+    abstract fun buildVisitor(holder: ProblemsHolder): PsiElementVisitor
+
+    private fun isEnabled(holder: ProblemsHolder): Boolean {
+        val file = holder.file
+        if (file !is KtFile)
+            return false
+        return file.annotationEntries.any {
+            it.shortName.toString() == "Verik"
+        }
+    }
+
+    companion object {
+
+        private val NULL_VISITOR = object : KtVisitorVoid() {}
     }
 }