+++ src/main/kotlin/io/verik/intellij/inspection/common/VerikInspectionProvider.kt
@@ -18,6 +18,7 @@ package io.verik.intellij.inspection.common
 
 import com.intellij.codeInspection.InspectionToolProvider
 import com.intellij.codeInspection.LocalInspectionTool
+import io.verik.intellij.inspection.inspection.TopNotModuleInspection
 import io.verik.intellij.inspection.inspection.UnsupportedElementInspection
 import io.verik.intellij.inspection.inspection.UnsupportedModifierInspection
 
@@ -26,7 +27,8 @@ class VerikInspectionProvider : InspectionToolProvider {
     override fun getInspectionClasses(): Array<Class<out LocalInspectionTool>> {
         return arrayOf(
             UnsupportedElementInspection::class.java,
-            UnsupportedModifierInspection::class.java
+            UnsupportedModifierInspection::class.java,
+            TopNotModuleInspection::class.java
         )
     }
 }