+++ src/main/kotlin/io/verik/intellij/inspection/inspection/UnsupportedElementInspection.kt
@@ -20,7 +20,6 @@ import com.intellij.codeHighlighting.HighlightDisplayLevel
 import com.intellij.codeInspection.ProblemsHolder
 import com.intellij.psi.PsiElementVisitor
 import io.verik.intellij.inspection.common.AbstractVerikInspection
-import io.verik.intellij.inspection.common.InspectionUtil
 import org.jetbrains.kotlin.psi.KtClassLiteralExpression
 import org.jetbrains.kotlin.psi.KtDestructuringDeclaration
 import org.jetbrains.kotlin.psi.KtDoubleColonExpression
@@ -51,10 +50,8 @@ class UnsupportedElementInspection : AbstractVerikInspection() {
         return HighlightDisplayLevel.ERROR
     }
 
-    override fun buildVisitor(holder: ProblemsHolder, isOnTheFly: Boolean): PsiElementVisitor {
-        return if (InspectionUtil.isEnabled(holder)) {
-            UnsupportedElementVisitor(holder)
-        } else InspectionUtil.NULL_VISITOR
+    override fun buildVisitor(holder: ProblemsHolder): PsiElementVisitor {
+        return UnsupportedElementVisitor(holder)
     }
 
     private class UnsupportedElementVisitor(